Split vis geometry and plot tests into module-owned files
vis_geometry_test.py, vis_plot_test.py, and vis_real_files_test.py were
catch-alls: between them they held the only coverage of scene2d.py,
scene3d.py, plot2d.py, plot3d.py, backend.py, backend_matplotlib.py,
compare2d.py, and five of the six layout families, none of which had a
sibling test file. Refile every test under the module whose API it calls.

New: scene2d_test.py, scene3d_test.py, plot2d_test.py, backend_test.py,
backend_matplotlib_test.py, compare2d_test.py. Appended to existing
plot3d_test.py, config_test.py, _values_test.py, and the six
layout/_*_test.py files.

Class names are distinct per destination so the three-way seeding of
scene3d_test.py (BuildRenderScene3dTest / RealFileScene3dTest /
Scene3dOverlayTest) does not collide.

vis/_testing.py gains VisDefaultsResetMixin, which replaces the
reset_defaults() + addCleanup boilerplate every split-out class needs.

config_test.py keeps both restore-on-exception tests: the existing one
asserts matplotlib rcParams are restored by publication_theme(), the
incoming one asserts branch_type_colors is restored by vis.theme().
Different subjects, not a duplicate.

vis/ collects 316 tests, down 34 from 350 — exactly the count
vis_plot_test.py contributed, so nothing was dropped.
