+++ braincell/_compute/layouts_test.py
@@ -735,7 +735,6 @@ class FnFingerprintWarnsOnOpaqueClosureTest(unittest.TestCase):
         )
 
 
-
 # =============================================================================
 # Spatially-varying (callable) mechanism parameters
 # =============================================================================
