+++ braincell/vis/layout/_fan_test.py
@@ -21,7 +21,7 @@ import numpy as np
 
 from braincell import Branch, Morphology
 from braincell.vis._testing import make_fan_root_partition_tree, make_two_dendrite_tree
-from braincell.vis.layout import point_on_layout_branch
+from braincell.vis.layout import build_layout_branches_2d, point_on_layout_branch
 from braincell.vis.layout._fan import _build_layout_branches_fan
 from braincell.vis.layout._common import _build_layout_specs
 
@@ -129,6 +129,62 @@ class BuildFanLayoutTest(unittest.TestCase):
             )
         )
         self.assertGreaterEqual(math.degrees(math.acos(cos_theta)), 34.0)
+
+
+class FanLayoutPartitionTest(unittest.TestCase):
+    def test_fan_root_parent_x_partition_uses_left_middle_right_sectors(self) -> None:
+        tree = make_fan_root_partition_tree()
+
+        layouts = {
+            layout.branch_name: layout
+            for layout in build_layout_branches_2d(tree, mode="tree", layout_family="fan")
+            if layout.branch_name != "soma"
+        }
+
+        self.assertLess(layouts["left_dend"].end_direction_um[0], 0.0)
+        self.assertGreater(layouts["mid_dend"].end_direction_um[1], 0.0)
+        self.assertLess(layouts["mid_axon"].end_direction_um[1], 0.0)
+        self.assertGreater(layouts["right_near"].end_direction_um[0], 0.0)
+        self.assertGreater(layouts["right_far"].end_direction_um[0], 0.0)
+
+    def test_fan_respects_min_branch_angle_for_siblings(self) -> None:
+        tree = make_two_dendrite_tree()
+        layouts = {
+            layout.branch_name: layout
+            for layout in build_layout_branches_2d(
+                tree,
+                mode="tree",
+                layout_family="fan",
+                min_branch_angle_deg=30.0,
+            )
+            if layout.branch_name.startswith("dend")
+        }
+        angle = math.degrees(
+            math.acos(
+                float(
+                    np.clip(
+                        np.dot(layouts["dend_a"].end_direction_um, layouts["dend_b"].end_direction_um)
+                        / (
+                            np.linalg.norm(layouts["dend_a"].end_direction_um)
+                            * np.linalg.norm(layouts["dend_b"].end_direction_um)
+                        ),
+                        -1.0,
+                        1.0,
+                    )
+                )
+            )
+        )
+        self.assertGreaterEqual(angle, 29.0)
+
+    def test_fan_branches_keep_constant_direction(self) -> None:
+        tree = make_fan_root_partition_tree()
+        layouts = [
+            layout
+            for layout in build_layout_branches_2d(tree, mode="tree", layout_family="fan")
+            if layout.branch_name != "soma"
+        ]
+        for layout in layouts:
+            self.assertTrue(np.allclose(layout.segment_directions_um[0], layout.segment_directions_um[-1]))
 
 
 if __name__ == "__main__":
