+++ braincell/vis/layout/_legacy_test.py
@@ -14,12 +14,14 @@
 # ==============================================================================
 
 
+import math
 import unittest
 import warnings
 
 import numpy as np
 
 from braincell.vis._testing import make_two_dendrite_tree
+from braincell.vis.layout import build_layout_branches_2d
 from braincell.vis.layout._common import _build_layout_specs
 from braincell.vis.layout._legacy import _build_layout_branches_legacy
 
@@ -38,6 +40,27 @@ class BuildLegacyLayoutTest(unittest.TestCase):
         self.assertEqual(len(layouts), len(tree.branches))
         for layout in layouts:
             self.assertTrue(np.all(np.isfinite(layout.segment_points_um)))
+
+
+class LegacyRootLayoutAngleTest(unittest.TestCase):
+    def test_min_branch_angle_deg_is_applied_in_legacy_layout(self) -> None:
+        tree = make_two_dendrite_tree()
+
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", DeprecationWarning)
+            layouts = build_layout_branches_2d(
+                tree,
+                mode="tree",
+                min_branch_angle_deg=90.0,
+                root_layout="legacy",
+            )
+        child_angles = sorted(
+            math.degrees(math.atan2(layout.end_direction_um[1], layout.end_direction_um[0]))
+            for layout in layouts
+            if layout.branch_name in {"dend_a", "dend_b"}
+        )
+
+        self.assertGreaterEqual(child_angles[1] - child_angles[0], 90.0 - 1e-6)
 
 
 if __name__ == "__main__":
